app/src/main/java/com/example/tuicodewars/data/utils/NetworkCheckerImpl.kt
package com.example.tuicodewars.data.utils

import android.content.Context
import android.net.ConnectivityManager
import android.net.NetworkCapabilities
import com.example.tuicodewars.domain.utils.NetworkChecker
import javax.inject.Inject

class NetworkCheckerImpl @Inject constructor(private val context: Context) : NetworkChecker {
    override fun isInternetAvailable(): Boolean {
        val connectivityManager =
            context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
        val network = connectivityManager.activeNetwork ?: return false
        val networkCapabilities =
            connectivityManager.getNetworkCapabilities(network) ?: return false
        return networkCapabilities.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
    }
}
